[refactor] Major API overhaul
