+++ .gitignore
@@ -2,3 +2,4 @@ __pycache__
 /dist
 /.vscode
 /.pytest_*
+**.egg-info
