--- mlinstruct/evaluate/metrics/classification/__init__.py
@@ -1,4 +1,0 @@
-from .confusion_matrix import ConfusionMatrix
-from .roc import ROC
-
-__all__ = ["ConfusionMatrix", "ROC"]
