+++ pyproject.toml
@@ -3,16 +3,21 @@ name = "mlinstruct"
 version = "0.1.0"
 description = "Python library for managing ML model training instances and evaluation graph generation"
 readme = "README.md"
-authors = [{ name = "Rahul Pavithran", email = "rahul.pavithran99@gmail.com" }]
+authors = [{ name = "Rahul Pavithran", email = "rahul.pavi@outlook.com" }]
 license = { text = "MIT" }
-requires-python = "<4.0,>=3.11"
+requires-python = "<4.0,>=3.12"
 dependencies = [
-    "numpy<2.0.0,>=1.26.2",
-    "matplotlib<4.0.0,>=3.6.3",
-    "torch<3.0.0,>=2.4.0",
-    "torchvision<1.0.0,>=0.19.0",
-    "scikit-learn<2.0.0,>=1.5.1",
-    "torchinfo<2.0.0,>=1.8.0",
+    "numpy>=1.26.2,<2.0.0",
+    "matplotlib>=3.6.3,<4.0.0",
+    "scikit-learn>=1.5.1,<2.0.0",
+    "markupsafe>=2.1.5,<3.0.0",
+]
+
+[project.optional-dependencies]
+torch = [
+    "torch>=2.4.0,<2.6.0",
+    "torchvision>=0.15.0,<1.0.0",
+    "torchinfo>=1.8.0,<2.0.0",
 ]
 
 [project.urls]
@@ -22,19 +27,21 @@ repository = "https://github.com/alpha2303/mlinstruct"
 test = ["pytest<9.0.0,>=8.1.1", "pytest-mock<4.0.0,>=3.14.0"]
 
 [[tool.uv.index]]
-name = "pytorch-cu124"
-url = "https://download.pytorch.org/whl/cu124"
+name = "pytorch-cpu"
+url = "https://download.pytorch.org/whl/cpu"
+
+[[tool.uv.index]]
+name = "pytorch-cu126"
+url = "https://download.pytorch.org/whl/cu126"
 explicit = true
 
-[tool.uv.sources]
-torch = [
-    { index = "pytorch-cpu" },
-    { index = "pytorch-cu124", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
-]
-torchvision = [
-    { index = "pytorch-cpu" },
-    { index = "pytorch-cu124", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
-]
+[build-system]
+requires = ["setuptools>=61.0"]
+build-backend = "setuptools.build_meta"
+
+[tool.setuptools]
+package-dir = { "" = "src" }
+packages = ["mlinstruct"]
 
 [tool.pytest.ini_options]
 pythonpath = ["."]
