+++ src/mlinstruct/__init__.py
@@ -1,0 +1,7 @@
+__version__ = "0.1.0"
+
+from . import evaluate
+from . import train
+from . import utils
+
+__all__ = ["evaluate", "train", "utils"]
