+++ src/mlinstruct/evaluate/__init__.py
@@ -1,0 +1,4 @@
+from . import metrics
+from . import plots
+
+__all__ = ["metrics", "plots"]
