+++ src/mlinstruct/evaluate/metrics/__init__.py
@@ -1,0 +1,4 @@
+from . import classification
+from .metric_utils import MetricUtils
+
+__all__ = ["classification", "MetricUtils"]
