+++ src/mlinstruct/evaluate/plots/__init__.py
@@ -1,0 +1,7 @@
+from .base_plotter import BasePlotter
+from .roc_plotter import ROCPlotter
+from .loss_plotter import LossPlotter
+from .cm_plotter import ConfusionMatrixPlotter
+
+__all__ = ["BasePlotter", "ROCPlotter", "LossPlotter", "ConfusionMatrixPlotter"]
+
