+++ src/mlinstruct/train/data_payload/__init__.py
@@ -1,0 +1,4 @@
+from .base_data_payload import BaseDataPayload
+from .torch_data_payload import TorchDataPayload
+
+__all__ = ["BaseDataPayload", "TorchDataPayload"]
