+++ src/mlinstruct/train/model_proxy/__init__.py
@@ -1,0 +1,4 @@
+from .base_model_proxy import BaseModelProxy
+from .torch_model_proxy import TorchModelProxy
+
+__all__ = ["BaseModelProxy", "TorchModelProxy"]
