+++ src/mlinstruct/utils/__init__.py
@@ -1,0 +1,4 @@
+from . import exception
+from . import funcs
+
+__all__ = ["exception", "funcs"]
